cast to string before converting to R

+++ tests/testthat/test-dataframe-frame.R
@@ -732,7 +732,7 @@ test_that("unstack() works", {
 })
 
 test_that("$glimpse() works", {
-  df <- as_polars_df(iris)$with_columns(pl$lit(42)$cast(pl$Int8))
+  df <- as_polars_df(iris)$with_columns(pl$lit(42)$cast(pl$Int64))
   expect_snapshot(df$glimpse())
   expect_snapshot(df$glimpse(max_items_per_column = 2))
   expect_snapshot(df$glimpse(max_colname_length = 2))
